Fix loading of environment specific supported token lists
The token lists were loaded twice over. First load was as part of the js
bundle that contains all the code, all the token lists are baked into it.
Then the token list for the current environment was loaded again through
using a fetch request when the environment was set.

Now the token lists are all loaded once from the js bundle on initial
load, and then the appropriate list is selected when the user selects
a network + wallet. This makes it possible to pick the right list for
BSC vs Ethereum.