Add ability to connect wallets to Binance Smart Chain (#1408)
* Disable listening to approval and deposit events on Ethereum/BSC

Nothing was actually being done with any of these events.

* Change labels from Ethereum -> Binance Smart Chain based on network

* Get Ethereum/BSC block explorer API base URL from config

* Make things work without a generic gateway

Initial BSC deployment will only support LOOM deposit/withdrawal, so
only the LOOM gateway will be available.

* Support Binance Chain Wallet (browser extension)

* Change more labels from Ethereum -> Binance Smart Chain based on network

* Show which network the wallet is connected to in the header

* Update the wrong network detection to work with Binance Chain wallet

* Make detection of multiple wallet browser extensions more reliable

* Enable WalletConnect for BSC

Also add some tooltips to the BSC wallet list so users can figure
out which one they should use.

* Eliminate clunky browser extension detection

Instead of checking if wallet browser extensions are enabled on page
load check when the user actually attempts to use them.

* Fix loading of environment specific supported token lists

The token lists were loaded twice over. First load was as part of the js
bundle that contains all the code, all the token lists are baked into it.
Then the token list for the current environment was loaded again through
using a fetch request when the environment was set.

Now the token lists are all loaded once from the js bundle on initial
load, and then the appropriate list is selected when the user selects
a network + wallet. This makes it possible to pick the right list for
BSC vs Ethereum.

* Connect to the bsc-gateway Go contract when wallet is connected to BSC

* Add BSC Testnet config for asia1