Get WalletConnect working with Binance Smart Chain
WalletConnect was breaking the build, yarn couldn't install the package,
perhaps it was just too old. Bumped up the package to the latest version
to fix the build... and then had to make some tweaks to get everything
working again.

Also removed support for fortmatic and portis wallets since that was
never shipped into production, and having unused dependencies around
is just a liability.