Reload when a different environment is selected on the login page (#1429)
Switching the configuration on the login page did not have an immediate effect on some settings (`features` section for instance). Since changing configs is only supported on dev/staging dashboards it's easier to work around this issue by simply reloading the page when a new config is selected from the drop-down, that way the right config is loaded when the page reloads.