Fix hash signing with Binance Chain wallet (#1421)
The Binance Chain wallet API mimics MetaMask but has some quirks with the `personal_sign` web3 function we rely on for account mappings, specifically it doesn't implement `personal_sign` and instead provides a custom `bnbSign` function on the wallet instance. The problem is loom-js relies on an ethers.js signer that wraps the web3 provider, ethers.js doesn't know anything about this `bnbSign` function so it doesn't work with Binance Chain wallet. The quickest workaround is to monkey patch the ethers.js signer to use `bnbSign` instead of `personal_sign`, this is probably going to break if we upgrade ethers.js.

There are also some other quirks with the Binance Chain wallet, it seems to attempt to subscribe to `newBlockHeaders` events in order to determine if a tx has been confirmed, even though the wallet is not connected to a websocket endpoint. This probably warrants further investigation since this seems like a pretty major flaw that someone was bound to notice already, but for now we just swallow these subscription errors and wait some time before assuming the tx has been confirmed.